use matrix to get the invert of cell
