+++ batoms/plugins/highlight/highlight.py
@@ -9,7 +9,7 @@ import numpy as np
 from batoms.base.object import ObjectGN
 from batoms.plugins.base import PluginObject
 from .setting import HighlightSettings
-from batoms.utils.butils import get_node_by_name
+from batoms.utils.utils_node import get_node_by_name
 from batoms.utils import string2Number
 import logging
